requirements.txt
aiohttp==3.7.4.post0
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
base45==0.4.2
bcrypt==3.2.0
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
bsddb3==6.2.9
cbor2==5.4.1
certifi==2019.11.28
certvalidator==0.11.1
cffi==1.14.5
chardet==3.0.4
chromedriver==2.24.1
cliapp==1.20180812.1
click==8.0.1
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
configparser==5.0.2
cose==0.9.dev8
crayons==0.4.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dateparser==1.0.0
dbus-python==1.2.16
debugpy==1.4.1
decorator==5.0.9
defer==1.0.6
discord.py==1.7.3
distro==1.4.0
dnspython==2.1.0
ecdsa==0.17.0
entrypoints==0.3
f.lux-indicator-applet==1.2.1rc0
Flask==2.0.1
future==0.18.2
grpcio==1.16.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
impacket==0.9.22
imutils==0.5.4
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==7.0.2
jupyter-core==4.7.1
keyring==18.0.1
kiwisolver==1.3.2
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
ldap3==2.9
ldapdomaindump==0.9.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
multidict==5.1.0
nemo-emblems==5.0.0
nest-asyncio==1.5.1
netaddr==0.7.19
netifaces==0.10.4
nose==1.3.7
numpy==1.17.4
oauthlib==3.1.0
onboard==1.4.1
oscrypto==1.2.1
output==1.0.1
packaging==20.3
PAM==0.4.2
pandas==1.3.3
paramiko==2.7.2
parso==0.8.2
pefile==2021.5.24
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
ply==3.11
prompt-toolkit==3.0.20
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pycryptodomex==3.10.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==2.1.0
pymacaroons==0.13.0
pymssql==2.2.1
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.6
pyparted==3.11.2
pyRFC3339==1.1
pysmi==0.3.4
pysnmp==4.4.6
python-apt==2.0.0+ubuntu0.20.4.6
python-binance==1.0.12
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.0
python-kucoin==2.1.2
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.1.0
qrcode==6.1
regex==2021.7.6
reportlab==3.5.34
requests==2.21.0
requests-file==1.4.3
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
speedtest-cli==2.1.2
ssh-import-id==5.10
systemd-python==234
tinycss2==1.0.2
tldextract==2.2.1
toml==0.10.2
tornado==6.1
traitlets==5.1.0
ttystatus==0.38
twilio==7.0.0
typing-extensions==3.10.0.0
tzlocal==2.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.2
Unidecode==1.1.1
urllib3==1.24.3
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.1
xkit==0.0.0
yarl==1.6.3
youtube-dl==2021.4.26
zeroconf==0.24.4
